feat: Completed!
Service examples working also with test for the core of the system. There are no test cases for examples, they just log out that they are doing. I have confirmed releaseing of resources via inspection in Observatory. Unfortuneatly I can not come up with a nice easy way to do this in a test. Merging this will close
#2 Close #1 Close.